Broken Unit tests added. issues with 64 and 32 bit.
supressed warning with manual cast.

Debug code.

+++ Code/src/NncToTs.cpp
@@ -109,8 +109,8 @@ void NncToTs::getNncSoundData() {
 		rightVolumes = new float[sources];
 		for (int i = 0; i < sources; i++) {
 			//TODO (Ryan Lavin): the comments of left and right volume indicate that the following statements are backwards. I don't believe them. Check this -6/13/2016
-			leftVolumes[i] = selfPlayer->leftVolume(*audibleSources[i]);
-			rightVolumes[i] = selfPlayer->rightVolume(*audibleSources[i]);
+			leftVolumes[i] = (float)selfPlayer->leftVolume(*audibleSources[i]);
+			rightVolumes[i] = (float)selfPlayer->rightVolume(*audibleSources[i]);
 		}
 
 		//Find distortions of each source
